fix(dsp/io): preserve input dtype across 3 remaining promotion sites (#3749, #3752, #3760)
Three remaining float32 → float64 silent-promotion sites flagged by
the engine audit. Same drift class as already-fixed
#3658 / #3659 / #2450 / #3744 / #3748.

#3752 — `AdaptiveLimiter._oversample`: `np.zeros(len(audio) * factor)`
defaulted to float64; pass `dtype=audio.dtype` to both `np.zeros`
calls and to the kernel `np.ones`. Cast the convolve result back via
`.astype(input_dtype, copy=False)` at the return so any scipy/numpy
promotion inside the filter loop is undone.

#3760 — `ParallelBandProcessor._process_with_threads`: `np.sum(band_results,
axis=0)` promotes to the highest dtype seen across workers. A single
float64 result (e.g. from scipy `sosfiltfilt`) promoted the whole
output. Cast the sum back to `audio.dtype` at the return.

#3749 — `unified_loader.load_audio` normalize branch:
`audio_data / peak * 0.98` could promote under older NumPy rules
since `peak` was a float64 scalar. Cast `peak` and `0.98` to the
audio dtype before the divide/multiply so the chain stays in the
input dtype regardless of NumPy version.

19 of the io / limiter / parallel-band / processor tests pass with
no new failures.

CONSISTENCY: all three are the canonical dtype-drift pattern that
the engine audit grouped under one "preserve_dtype" sweep. Remaining
matches of `np.zeros(...)` / `np.sum(...)` in the codebase already
pass `dtype=` or pre-allocate `np.zeros_like(...)`.

+++ auralis/optimization/parallel_processor.py
@@ -312,8 +312,13 @@ class ParallelBandProcessor:
                     except Exception as exc:
                         warning(f"Band {band_i} processing failed (thread pool), using unprocessed signal: {exc}")
 
-        # Sum all band results
-        output: np.ndarray = np.sum(band_results, axis=0)
+        # #3760: sum all band results, preserving the input dtype.
+        # `np.sum` promotes to the highest dtype seen across the worker
+        # results, so a single float64 entry (e.g. from scipy's
+        # `sosfiltfilt`) promotes the entire output to float64 — same
+        # dtype-drift class as #3658 / #3659 / #2450 / #3752. Cast back
+        # explicitly.
+        output: np.ndarray = np.sum(band_results, axis=0).astype(audio.dtype, copy=False)
 
         return output
